requirements.txt
aiobotocore==1.0.7
amqp==2.5.2
asn1crypto==0.24.0
Babel==2.6.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.12.32
botocore==1.15.32
celery==4.4.1
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
dateparser==0.7.1
dj-rest-auth==2.1.3
Django==3.1
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap-static==4.2.1
django-choices==1.7.1
django-codemirror-widget==0.5.0
django-compressor==2.4
django-constance[redis]==2.8.0
django-configurations==2.2
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-filebrowser-no-grappelli==3.8.0
django-formtools==2.2
django-htmlmin==0.11.0
django-inline-svg==0.1.1
django-ipware==2.1.0
django-js-asset==1.2.2
django-jsoneditor==0.1.4
django-meta==1.7.0
django-mptt==0.11.0
django-nested-admin==3.1.3
django-partial-index==0.5.2
django-phonenumber-field==4.0.0
django-redis==4.12.1
django-smart-selects==1.5.9
django-storages==1.9.1
django-tools==0.45.3
django-user-agents==0.4.0
djangoajax==3.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
django-webpack-loader==0.7.0
docutils==0.14
easy-thumbnails==2.6
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
kombu==4.6.8
MarkupSafe==1.1.1
maxminddb==1.5.2
numpy==1.16.3
packaging==19.0
pbr==5.1.2
phonenumbers==8.10.22
Pillow==7.1.1
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.7.5
python-monkey-business==1.0.0
pytz==2018.7
rcssmin==1.0.6
redis==3.2.1
regex==2019.4.14
requests==2.21.0
rjsmin==1.1.0
s3transfer==0.3
selenium==3.141.0
sqlparse==0.2.4
suds-py3==1.3.3.0
text-unidecode==1.3
tornado==5.1.1
twilio==6.25.2
tzlocal==1.5.1
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.24.1
user-agents==2.0
vine==1.3.0
tinify==1.5.1
psutil==5.6.3
django-sortedm2m==3.0.2
pyopenssl==19.0.0
elasticsearch==7.0.5
elasticsearch-dsl==7.1.0
Markdown==3.1.1
aiohttp==3.6.2
django-extensions==3.0.4
sendgrid==6.2.0
celery-singleton==0.2.0
django-admin-shortcuts==2.0.0
django-admin-interface==0.13.4
social-auth-app-django==3.1.0
social-auth-core==3.3.3
django-cacheops==5.0
drf-yasg==1.17.1
six==1.15.0
